+++ apps/web/src/app/auth/callback/page.tsx
@@ -1,39 +1,7 @@
-"use client";
-export const dynamic = "force-dynamic";
-
-import { useEffect } from "react";
-import { useRouter, useSearchParams } from "next/navigation";
-import { supabase } from "@/shared/api/supabase/client";
+import { Suspense } from "react";
+import AuthCallbackClient from "./AuthCallbackClient";
 
 export default function AuthCallbackPage() {
-  const router = useRouter();
-  const searchParams = useSearchParams();
-  const fromExtension = searchParams.get("from") === "extension";
-
-  useEffect(() => {
-    const redirect = (session: boolean) => {
-      if (!session) {
-        router.replace("/login");
-        return;
-      }
-      // extension에서 온 경우 토큰 전달 페이지로
-      router.replace(fromExtension ? "/auth/extension-token" : "/");
-    };
-
-    const {
-      data: { subscription },
-    } = supabase.auth.onAuthStateChange((event, session) => {
-      if (event === "SIGNED_IN") redirect(!!session);
-      if (event === "SIGNED_OUT") router.replace("/login");
-    });
-
-    supabase.auth.getSession().then(({ data: { session } }) => {
-      if (session) redirect(true);
-    });
-
-    return () => subscription.unsubscribe();
-  }, [router, fromExtension]);
-
   return (
     <div className="flex min-h-screen items-center justify-center">
       <div className="flex flex-col items-center gap-3">
@@ -54,6 +22,9 @@ export default function AuthCallbackPage() {
         </svg>
         <p className="text-sm text-zinc-500">로그인 처리 중...</p>
       </div>
+      <Suspense>
+        <AuthCallbackClient />
+      </Suspense>
     </div>
   );
 }
